refactor: extract stream_series.go
